fixed small typo in readme ('test' -> 'text')
Fixed a small typo. Said "test" when it should have said "text":

"Note: When you hit the pry, you may end up with a prompt that looks like this `:`. That just means there's a lot of **test** [now **text**] and you can scroll up and down using the arrow keys"